backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "anyio>=4.12.1",
    "arq>=0.27.0",
    "click>=8.3.1",
    "colorama>=0.4.6",
    "fastapi>=0.135.1",
    "h11>=0.16.0",
    "httpx>=0.28.1",
    "idna>=3.11",
    "bcrypt<4.0.0",
    "mypy>=1.19.1",
    "passlib[bcrypt]>=1.7.4",
    "pgvector>=0.4.2",
    "pydantic>=2.12.5",
    "pyjwt>=2.11.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "redis>=5.3.1",
    "ruff>=0.15.5",
    "sqlalchemy>=2.0.48",
    "sqlmodel>=0.0.37",
    "starlette>=0.52.1",
    "typing-extensions>=4.15.0",
    "typing-inspection>=0.4.2",
    "uvicorn>=0.41.0",
    "pydantic-settings>=2.13.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]

